One component per file for essentia window.
